Check if mask and input have the same spacing to compute SUVPeak
Before, the output of the SUVPeak was unrealistic with a mask with a spacing different from the input.
Now, check if they have the same spacing. If not, the script raises an error.

I add an option (--allow_resize) to automatically resize the mask according to the input like in the clitkImageStatistics script