Change name of the button in Image Arithm tool
Before, it was QDialogButtonBox::Discard, but according to https://doc.qt.io/archives/qt-4.8/qdialogbuttonbox.html#StandardButton-enum
it can be displayed "Don't save".
For this tool, we want to select 1 image instead of 2, so the name of the button is not appropriate.
Moreover in skip function, the comparision to "Discard" can be complicated with another name.

So I prefer to use Ignore button because it's more appropriate